fix return value of _printf

+++ main.h
@@ -3,7 +3,7 @@
 
 int _printf(const char *format, ...);
 int _putchar(char c);
-int print_number(int *pos, int number);
+int print_number(int *pos, long int number);
 int num_to_char(int num);
 int print_binary(int *pos, unsigned int number);
 #endif
